remove useless consts & rename helper
